Rename to constants.py
